+++ src/main/resources/application.properties
@@ -2,7 +2,7 @@ quarkus.kubernetes.deployment-target=kubernetes, openshift
 quarkus.native.resources.includes=k8s/def/*.yaml
 quarkus.container-image.registry=quay.io
 quarkus.container-image.group=windupeng
-quarkus.container-image.name=${quarkus.application.name}-native
+quarkus.container-image.name=windup-operator-native
 quarkus.container-image.tag=latest
 quarkus.kubernetes.service-type=load-balancer
